pyproject.toml
[tool.poetry]
name = "livraria"
version = "0.1.0"
description = ""
authors = ["joao vitor g de luca"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.13.0"
gunicorn = "^20.1.0"
whitenoise = "^6.2.0"
djangorestframework-simplejwt = "^5.2.2"
python-magic = "^0.4.27"
pillow = "^9.3.0"
drf-spectacular = "^0.24.2"
django-environ = "^0.9.0"
mysqlclient = "^2.1.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
isort = "^5.10.1"
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
